refactor(api): Move custom rules structs and method to dedicated file
