+++ api/model/profile_test.go
@@ -89,64 +89,3 @@ func TestExportedCustomRuleGroups_MaxMatchesCanonicalConst(t *testing.T) {
 			field, got, ExportedCustomRuleGroupsLimit)
 	}
 }
-
-// names is a helper to pluck the ordered group names out of a list for assertions.
-func names(list []CustomRuleGroup) []string {
-	out := make([]string, len(list))
-	for i, g := range list {
-		out[i] = g.Name
-	}
-	return out
-}
-
-// TestCustomRuleGroups_Reorder exercises the display-order rewrite: exact reorder,
-// unknown names ignored, omitted groups appended in place, comment preservation,
-// per-list isolation, and no-op on an empty list.
-// tableRef: I8
-func TestCustomRuleGroups_Reorder(t *testing.T) {
-	t.Run("reorders to match and preserves comments", func(t *testing.T) {
-		g := CustomRuleGroups{Block: []CustomRuleGroup{
-			{Name: "Ads", Comment: "a"}, {Name: "Social", Comment: "s"}, {Name: "Work", Comment: "w"},
-		}}
-		g.Reorder(ACTION_BLOCK, []string{"Work", "Ads", "Social"})
-		assert.Equal(t, []string{"Work", "Ads", "Social"}, names(g.Block))
-		assert.Equal(t, "w", g.Block[0].Comment, "comment must travel with the group")
-	})
-
-	t.Run("ignores unknown names", func(t *testing.T) {
-		g := CustomRuleGroups{Block: []CustomRuleGroup{{Name: "Ads"}, {Name: "Work"}}}
-		g.Reorder(ACTION_BLOCK, []string{"Work", "Ghost", "Ads"})
-		assert.Equal(t, []string{"Work", "Ads"}, names(g.Block))
-	})
-
-	t.Run("appends omitted groups in their existing order", func(t *testing.T) {
-		g := CustomRuleGroups{Block: []CustomRuleGroup{
-			{Name: "Ads"}, {Name: "Social"}, {Name: "Work"},
-		}}
-		// Client only names one group; the rest keep their relative order at the end.
-		g.Reorder(ACTION_BLOCK, []string{"Work"})
-		assert.Equal(t, []string{"Work", "Ads", "Social"}, names(g.Block))
-	})
-
-	t.Run("ignores duplicate names in the order list", func(t *testing.T) {
-		g := CustomRuleGroups{Block: []CustomRuleGroup{{Name: "Ads"}, {Name: "Work"}}}
-		g.Reorder(ACTION_BLOCK, []string{"Work", "Work", "Ads"})
-		assert.Equal(t, []string{"Work", "Ads"}, names(g.Block))
-	})
-
-	t.Run("is per-list: reordering block leaves allow untouched", func(t *testing.T) {
-		g := CustomRuleGroups{
-			Block: []CustomRuleGroup{{Name: "Ads"}, {Name: "Work"}},
-			Allow: []CustomRuleGroup{{Name: "Ads"}, {Name: "Work"}},
-		}
-		g.Reorder(ACTION_BLOCK, []string{"Work", "Ads"})
-		assert.Equal(t, []string{"Work", "Ads"}, names(g.Block))
-		assert.Equal(t, []string{"Ads", "Work"}, names(g.Allow), "allow list must be unchanged")
-	})
-
-	t.Run("no-op on empty list", func(t *testing.T) {
-		g := CustomRuleGroups{}
-		g.Reorder(ACTION_BLOCK, []string{"Ghost"})
-		assert.Nil(t, g.Block)
-	})
-}
